% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/al_anc_query.R
\name{al_anc_query}
\alias{al_anc_query}
\title{Aluminum get ancillary data}
\usage{
al_anc_query(al_df)
}
\arguments{
\item{al_df}{data from AWQMS that contains al data. Ancillary data gets joined to this.}
}
\value{
data from AWQMS that contains al ancillary data.
}
\description{
This function queries AWQMS to get ancillary data neede dto calculate Al criteria
}
