% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Al_ancillary_params}
\alias{Al_ancillary_params}
\title{Al_ancillary_params}
\format{
A vector with 7 elements
\describe{
\item{Al_ancillary_params}{Characteristic name in AWQMS for needed parameters to calculate criteria}
}
}
\usage{
Al_ancillary_params
}
\description{
A vector of parameters needing to query to calculate Al criteria
}
\keyword{datasets}
