% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/al_combine_ancillary.R
\name{al_combine_ancillary}
\alias{al_combine_ancillary}
\title{Aluminum Combine Ancillary Data}
\usage{
al_combine_ancillary(al_df, ancillary_df)
}
\arguments{
\item{al_df}{data from AWQMS that contains al data. Ancillary data gets joined to this.}

\item{ancillary_df}{Output of al_get_ancillary()}
}
\value{
Dataframe with DOC, Hardness, pH, al_ancillary_cmt added to al_df
}
\description{
This function combines Al data and output from al_get_ancillary in preparation to calculate criteria
}
