% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aL_crit_calculator.R
\name{al_crit_calculator}
\alias{al_crit_calculator}
\title{Aluminum Criteria Calculator}
\usage{
al_crit_calculator(
  df,
  ph_col = "pH",
  hardness_col = "Hardness",
  DOC_col = "DOC",
  lat_col = "Lat_DD",
  long_col = "Long_DD",
  verbose = FALSE
)
}
\arguments{
\item{df}{Dataframe to modify}

\item{ph_col}{Name of column containing pH values. Must be surrounded by quotes}

\item{hardness_col}{Name of column containing hardness values. Must be surrounded by quotes}

\item{DOC_col}{Name of column containing DOC values. Must be surrounded by quotes}

\item{verbose}{If true, return genus, and other toxicity background information.}
}
\description{
This function calculates aluminum criteria based on EPA's 2018 national recommended freshwater aquatic life criteria
for aluminum. This function is based on the EPA provided function located here:
https://www.epa.gov/sites/production/files/2020-01/aluminum-criteria-calculator-rcode-data.zip
The dataframe needs to have columns for pH, hardness, and DOC values. This function adds 6 columns to original
dataframe:  CCC, FAV, CMC, Final_CMC, FINAL_CCC, Flag.
}
