% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/al_default_criteria.R
\name{al_default_criteria}
\alias{al_default_criteria}
\title{Aluminum default criteria Lookup}
\usage{
al_default_criteria(lat, long, type = c("Chronic"))
}
\arguments{
\item{lat}{Latitude of station}

\item{long}{Longitude of station}

\item{type}{If 'Chronic' return CCC, if 'Acute' return CMC}
}
\value{
default criteria value. If NA is returned, station is outside of georeferenced area.
}
\description{
This function looks up default DOC values used in calculating Al criteria. The function sends lat/long values via
a query to a Oregon DEQ map server to determine deafult DOC values. DOC values are based on georegion as described in
##insert link to interpretation doc when finalized##. If -9999 is returned, station is outside of georeferenced area.
}
\examples{
\dontrun{
Al_default_criteria(45.57770, -122.7476, type = "Acute")
Al_default_criteria(45.57770, -122.7476, type = "Chronic")

#Return chronic criteria
aluminum_data_mlocs <- aluminum_data \%>\%
 select(MLocID, Lat_DD, Long_DD) \%>\%
 distinct() \%>\%
 rowwise() \%>\%
 mutate(def_DOC = Al_default_criteria(Lat_DD, Long_DD))
}
}
