% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all}
\alias{all}
\title{Toxicology data needed to calculate Al criteria}
\format{
A dataframe with 151 rows and 13 variables
}
\source{
\url{https://www.epa.gov/sites/production/files/2020-01/aluminum-criteria-calculator-rcode-data.zip}
}
\usage{
all
}
\description{
A dataframe of 304(a) acute and chronic datasets prepared for processing. These are generated using lines 33-71 of
EPA supplied Al criteria calculator. It is awkwardly named all, beacuse that is what EPA named it.
}
\keyword{datasets}
